rename: components 폴더에서 Layouts 폴더 분리 [TRL-127]
