+++ src/Layouts/HeaderLayout/index.tsx
@@ -1,7 +1,7 @@
 import { Outlet } from 'react-router-dom';
 import styled from 'styled-components';
 
-import Header from '@/components/Layouts/HeaderLayout/Header';
+import Header from '@/components/Header';
 import { HEADER_HEIGHT } from '@/styles/constants';
 
 const HeaderLayout = () => {
